requirements.txt
python-telegraph
google-generativeai
telegraph